fix verification bug and login issue

+++ controllers/user.js
@@ -29,13 +29,11 @@ module.exports.createUser = async (req, res) => {
             html: mailVerification(token)
         })
         // console.log(newUser, newToken);
-        req.flash('success', 'verifikasi akun melalui email anda')
-        res.redirect('/user/verification')
-        // req.login(user, function(err) {
-        //     if (err) { return next(err); }
-        //     req.flash('success', 'sukses membuat user')
-        //     res.redirect(`/`)
-        // });
+        req.login(user, function(err) {
+            if (err) { return next(err); }
+            req.flash('success', 'sukses membuat user, masukan token pada email anda')
+            res.redirect('/user/verification')
+        });
     } catch (e) {
         req.flash('error', e.message, e.stack);
         res.redirect(`/user`);
@@ -67,10 +65,10 @@ module.exports.userVerificationPage = async (req, res) => {
 // cek kenapa masih pake try catch
 module.exports.userVerification = async (req, res) => {
     try{
-        const {username, token} = req.body;
-        const user = await User.findOne({username});
+        const {token} = req.body;
+        const user = req.user
         if(!user) {
-            req.flash('error', 'Token atau username salah');
+            req.flash('error', 'Harus Login terlebih dahulu');
             res.redirect(`/user/verification`);
         }
         const storedToken = await UserVerification.findOne({owner:user._id});
@@ -83,14 +81,13 @@ module.exports.userVerification = async (req, res) => {
             req.flash('error', 'Token atau username salah');
             res.redirect(`/user/verification`);
         }
-        user.verified = true;
-        await UserVerification.findByIdAndDelete(storedToken._id)
-        await user.save( )
-        req.login(user, function(err) {
-            if (err) { return next(err); }           
+        if (isMatched) {
+            user.verified = true;
+            await UserVerification.findByIdAndDelete(storedToken._id)
+            await user.save()
             req.flash('success', 'sukses membuat user')
             res.redirect(`/`)   
-        });
+        }
     } catch (e) {
         req.flash('error', e.message, e.stack);
         res.redirect(`/user`);
@@ -99,6 +96,7 @@ module.exports.userVerification = async (req, res) => {
 
 module.exports.resendToken = async (req, res) => {
     const user = req.user
+    await UserVerification.deleteMany({owner: user._id})
     // console.log(req.user) // req user itu ada di session
     const token = generateToken()
     const userVerification = new UserVerification({
@@ -112,7 +110,7 @@ module.exports.resendToken = async (req, res) => {
         subject: 'verify your account',
         html: mailVerification(token)
     })
-    console.log(user)
+    req.flash('success', 'token sudah dikirim ke email anda')
     res.redirect('/user/verification')
 }
